Ammended CMake file to work(?) with an "install"
Added a second pair of globs to find the required objects when they have
been "installed". Nektar++ does not have an install procedure for the
objects required to build ADRSolver and CompressibleFlowSolver. This
assumes that a spack package placed the solver objects in
PREFIX/solvers_objects/*.

Needs testing as I currently can't run the top level cmake file.