Added combined solver and poisson inputs from nektar++ (#59)
* Added CombinedSolver and cmake file.

Added bare-bones Nektar++ program and cmake file. Cmake file uses
existing nektar objects for ADRSolver and CompressibleFlowSolver.
These are passed to cmake with the NEKTAR_BUILD_PATH configuration
option.

* Added Nektar++ Poisson example.

Copied the Poisson example from nektar++. This should work with the
combined solver.

* README.rst -> README.md

* re-added revision.hpp but changed default git REVISION and STATUS strings

* this isnt perfect cos stil has a ..
but works with the compiler directive / env var that is currently passed in

* turn verbose off again

* error messages referred to old env var

* it renamed the main exe

* Ammended CMake file to work(?) with an "install"

Added a second pair of globs to find the required objects when they have
been "installed". Nektar++ does not have an install procedure for the
objects required to build ADRSolver and CompressibleFlowSolver. This
assumes that a spack package placed the solver objects in
PREFIX/solvers_objects/*.

Needs testing as I currently can't run the top level cmake file.

* delete include/revision.hpp again and add .fld and .opt to ignore

Co-authored-by: James Cook <james.cook@ukaea.uk>
Co-authored-by: James Cook <cookjws@gmail.com>