Fixed segfault in geometry test when building with DPC++/MKL
When building with the refactored build system, a segfault was
occurring in the test `GeometryTransport2DTest.QuadClone`. After quite
a lot of digging, this was determined to be the result of the order of
the libraries `libmkl_intel_ilp64.so` and `libmkl_intel_lp64.so` in
the link command. These present different [interfaces for
MKL](https://www.intel.com/content/www/us/en/develop/documentation/onemkl-linux-developer-guide/top/linking-your-application-with-onemkl/linking-in-detail/linking-with-interface-libraries/using-the-ilp64-interface-vs-lp64-interface.html),
with the former indexing arrays using 64-bit integers and the latter
using 32-bit integers. Nektar++ was built with the latter, but the
DPC++ version of MKL only supports the former. Mixing these two
interfaces seems risky...

The test can compile run successfully provided that
`libmkl_intel_lp64.so` appears **first** in the link command. Some
changes I'd made with my refactoring meant that it was not being
included at all, while `libmkl_intel_ilp64.so` still was. (The build
and test do work if neither of those libraries is present in the link
command, strangely). Even if it had been, the libraries would have
been in the wrong order. I have now corrected this.