Refactor cmake to use modern conventions (#118)
* Reorganise tests and put build products in right place

This addresses #102, #103, and #110.

* Removed duplicate find_package call for MPI

* Removed commented out configurations

* Applied cmake-format

* Initial attempts to move logic out of main CMakeLists.txt

* Created separate file for finding FFT implementation

Also simplified and tidied up the main CMake script quite a bit

* Modernised build script for Nektar solver

* Refactor dependencies and target definition for build

* Modernised installation of build products

* Created interface target for Nektar++

* Ensured build flags are actually used

* Try going back to old way of linking MKL

* Restore refactored version of MKL FFT library

* Fixed segfault in geometry test when building with DPC++/MKL

When building with the refactored build system, a segfault was
occurring in the test `GeometryTransport2DTest.QuadClone`. After quite
a lot of digging, this was determined to be the result of the order of
the libraries `libmkl_intel_ilp64.so` and `libmkl_intel_lp64.so` in
the link command. These present different [interfaces for
MKL](https://www.intel.com/content/www/us/en/develop/documentation/onemkl-linux-developer-guide/top/linking-your-application-with-onemkl/linking-in-detail/linking-with-interface-libraries/using-the-ilp64-interface-vs-lp64-interface.html),
with the former indexing arrays using 64-bit integers and the latter
using 32-bit integers. Nektar++ was built with the latter, but the
DPC++ version of MKL only supports the former. Mixing these two
interfaces seems risky...

The test can compile run successfully provided that
`libmkl_intel_lp64.so` appears **first** in the link command. Some
changes I'd made with my refactoring meant that it was not being
included at all, while `libmkl_intel_ilp64.so` still was. (The build
and test do work if neither of those libraries is present in the link
command, strangely). Even if it had been, the libraries would have
been in the wrong order. I have now corrected this.

* Applied cmake-format

* Clarified ctest versus gtest

* Made it optional to have pre-existing gtest installation (again)

* add spack install --test root to README.md

* Updated readme to reflect fixing of binary build directory

* Use latest version of NESO-spack repository

* Switch to linking MKL statically

Seems to help address the issue with using the 32-bit interface for
Nektar++ and the 64-bit interface for NESO.

* Always build tests with spack

* Removed .github/workflows/neso.yaml

* Remove reference to --test argument with spack

Co-authored-by: James Cook <cookjws@gmail.com>
Co-authored-by: Will Saunders <will.saunders@ukaea.uk>