Ensure CMake finds spack installations of MKL instead of system
CMake only seems to look for libraries under MKL_ROOT if that has been
set _outside_ of the config file (don't know why). This meant the
`find_library` command was searching in system paths before resorting
to the `PATH` it had been passed directing it to the local
directory. If a user already has a different installation of MKL on
their system, this can lead to link-time errors.