Improvements to development environment/workflow (#132)
* Spack environment now compiles both GCC and Intel builds

Also added a convenience script to activate the environment and set
some additional environment variables

* Updated readme to reflect change in workflow

* Update Spack repo with Owen's latest fixes

* Activate script creates symlinks with meaningful names to build dirs

* Made better messages in activate script and updated readme

* Updated gitignore to reflect changes to build directories, etc.

* Move spack repo to main

* Remove absolute paths to views

* Ignore dirty Nektar submodule

This submodule will always show as dirty due to the application of a
patch.

* Update Spack repo with DPCPP package improvements

* Ensure CMake finds spack installations of MKL instead of system

CMake only seems to look for libraries under MKL_ROOT if that has been
set _outside_ of the config file (don't know why). This meant the
`find_library` command was searching in system paths before resorting
to the `PATH` it had been passed directing it to the local
directory. If a user already has a different installation of MKL on
their system, this can lead to link-time errors.

* Ensure links to build directories always updated

* fixed unsetting of variables when deactivating environment

* Updated README to include note about submodules